refactor: reorder data module selectors and tests to prevent circular imports
