+++ app/javascript/modules/data/redux/selectors/dataSelectors.test.js
@@ -1,331 +1,365 @@
 import dotProp from 'dot-prop-immutable';
 import { DEFAULT_LOCALES } from 'modules/constants';
 import {
+    DEFAULT_MAP_SECTION,
     getCollectionsForCurrentProject,
     getCollectionsStatus,
     getContributionsStatus,
+    getContributionTypesForCurrentProject,
     getContributorsFetched,
+    getCurrentInterview,
     getCurrentInterviewFetched,
+    getCurrentProject,
     getCurrentRefTree,
     getCurrentRefTreeStatus,
     getCurrentUser,
     getCurrentUserIsAdmin,
     getFlattenedRefTree,
+    getGroupedContributions,
     getHeadings,
     getHeadingsFetched,
     getHeadingsStatus,
     getInterviewsStatus,
-    getLanguages,
+    getIsCampscapesProject,
+    getIsCatalog,
     getLanguagesStatus,
+    getMapSections,
     getMarkTextStatus,
     getMediaStreamsForCurrentProject,
     getPeopleStatus,
-    getPermissions,
     getPermissionsStatus,
     getProjectLocales,
     getProjectsStatus,
-    getRandomFeaturedInterviews,
+    getProjectTranslation,
     getRefTreeStatus,
-    getRegistryEntries,
     getRegistryEntriesStatus,
     getRegistryNameTypesForCurrentProject,
     getRolesForCurrentProject,
     getRolesStatus,
-    getSegments,
     getSegmentsStatus,
+    getShowFeaturedInterviews,
+    getShowStartPageVideo,
     getSpeakerDesignationsStatus,
-    getStatuses,
-    getTasks,
     getTasksStatus,
     getTaskTypesForCurrentProject,
     getTaskTypesStatus,
     getTranscriptFetched,
-    getUsers,
     getUsersStatus,
 } from './dataSelectors';
+import {
+    contributionState,
+    interviewer,
+    projectState,
+    qualityManager,
+    sound1,
+    sound2,
+    state,
+} from './stateFixture';
 
-const state = {
-    archive: {
-        archiveId: 'cd003',
-        projectId: 'cdoh',
-    },
-    data: {
-        users: {
-            current: {
-                id: 45,
-                admin: true,
-            },
+/** Interview Tests */
+
+test('getCurrentInterviewFetched retrieves if current interview has been fetched', () => {
+    expect(getCurrentInterviewFetched(state)).toBe(true);
+});
+
+test('getHeadings gets headings object of current interview', () => {
+    expect(getHeadings(state)).toEqual(state.data.interviews.cd003.headings);
+});
+
+describe('getTranscriptFetched', () => {
+    test('returns true if transcript has been fetched', () => {
+        expect(getTranscriptFetched(state)).toBeTruthy();
+    });
+
+    test('returns false if transcript has not been fetched', () => {
+        const _state = dotProp.set(state, 'archive.archiveId', 'za085');
+        expect(getTranscriptFetched(_state)).toBeFalsy();
+    });
+});
+
+test('getContributorsFetched retrieves if contributors for current interview have been fetched', () => {
+    expect(getContributorsFetched(state)).toBe(true);
+});
+
+test('getCurrentRefTree retrieves ref tree of current interview', () => {
+    expect(getCurrentRefTree(state)).toEqual(
+        state.data.interviews.cd003.ref_tree
+    );
+});
+
+test('getFlattenedRefTree retrieves flattened version of ref tree', () => {
+    const actual = getFlattenedRefTree(state);
+    const expected = {
+        11310: {
+            id: 11310,
+            type: 'node',
+            children: [
+                {
+                    type: 'leafe',
+                    time: 1472.18,
+                    tape_nbr: 1,
+                },
+                {
+                    type: 'leafe',
+                    time: 2103.32,
+                    tape_nbr: 1,
+                },
+            ],
         },
-        statuses: {
-            users: {
-                current: 'fetched',
-                resultPagesCount: 1,
-            },
-            collections: {
-                for_projects_1: 'fetched',
-            },
-            contributions: {
-                14: 'fetched-Mon',
-            },
-            headings: {
-                for_interviews_cd003: 'fetched-Mon',
-            },
-            interviews: {
-                za283: 'fetched',
-            },
-            languages: {
-                all: 'fetched',
-            },
-            mark_text: {
-                for_interviews_za003: 'fetched-Mon',
-            },
-            people: {
-                12: 'fetched',
-                contributors_for_interview_22: 'fetched',
-                for_projects_1: 'fetched',
-            },
-            permissions: {
-                page_1: 'fetched-Mon',
-            },
-            projects: {
-                all: 'fetched',
-            },
-            random_featured_interviews: {
-                all: 'fetched-Thu Dec 17 2020 18:20:22 GMT+0100 (Central European Standard Time)',
-            },
-            ref_tree: {
-                for_interviews_cd003: 'fetched-Mon',
-            },
-            registry_entries: {
-                1: 'fetched-Thu Jan 07 2021 21:17:39 GMT+0100 (Central European Standard Time)',
-            },
-            roles: {
-                page_1: 'fetched-Mon',
-            },
-            segments: {
-                for_interviews_za003: 'fetched-Mon',
-                for_interviews_cd003: 'fetched-Tue',
-            },
-            speaker_designations: {
-                for_interviews_za003: 'first_step_explanation',
-            },
-            tasks: {
-                35: 'processed',
-            },
-            task_types: {
-                for_projects_1: 'fetched',
-            },
+        11786: {
+            id: 11786,
+            type: 'node',
+            children: [
+                {
+                    type: 'leafe',
+                    time: 25.2,
+                    tape_nbr: 1,
+                },
+            ],
         },
-        projects: {
-            1: {
-                id: 1,
-                type: 'Project',
-                identifier: 'cdoh',
-                available_locales: ['de', 'en', 'ru'],
-                has_map: true,
-                root_registry_entry_id: 1,
-                collections: {
-                    1: {
-                        id: 1,
-                        type: 'Collection',
-                    },
-                },
-                people: {
-                    4: {
-                        id: 4,
-                        type: 'Person',
-                    },
-                },
-                registry_name_types: {
-                    2: {
-                        id: 2,
-                        type: 'RegistryNameType',
-                        code: 'first_name',
-                        name: 'Vorname',
-                    },
-                },
-                roles: {
-                    1: {
-                        id: 1,
-                        type: 'Role',
-                    },
-                },
-                task_types: {
-                    3: {
-                        id: 3,
-                        type: 'TaskType',
-                        key: 'protocol',
-                        abbreviation: 'Pro',
-                    },
-                },
-                contribution_types: {
-                    1: {
-                        id: 1,
-                        type: 'ContributionType',
-                        code: 'interviewee',
-                        project_id: 1,
-                        label: {
-                            de: 'Interviewte*r',
-                            en: 'Interviewee',
-                        },
-                    },
-                },
-                media_streams: {
-                    1: {
-                        id: 1,
-                        type: 'MediaStream',
-                        project_id: 1,
-                        path: 'https://medien.cedis.fu-berlin.de/zwar/zwar/#{archive_id}/#{archive_id}_#{tape_count}_0#{tape_number}_sd480p.mp4',
-                        media_type: 'video',
-                        resolution: '480p',
-                    },
-                },
+    };
+
+    expect(actual).toEqual(expected);
+});
+
+test('getCurrentInterview retrieves current interview', () => {
+    expect(getCurrentInterview(state)).toEqual(state.data.interviews.cd003);
+});
+
+describe('getGroupedContributions', () => {
+    test('returns contributions of current interview, ordered and grouped by type, without interviewee', () => {
+        const actual = getGroupedContributions(contributionState);
+        const expected = [
+            {
+                type: 'interviewer',
+                contributions: [interviewer],
             },
-        },
-        interviews: {
-            cd003: {
-                id: 22,
-                type: 'Interview',
-                lang: 'ru',
-                languages: ['ru', 'de'],
-                contributions: {
-                    1345: {
-                        id: 1345,
-                        type: 'Contribution',
-                        contribution_type: 'interviewee',
-                        person_id: 4,
-                        interview_id: 22,
-                    },
-                },
-                first_segments_ids: {
-                    1: 199498,
-                },
-                ref_tree: {
-                    id: 1,
-                    type: 'node',
-                    children: [
-                        {
-                            id: 11786,
-                            type: 'node',
-                            children: [
-                                {
-                                    type: 'leafe',
-                                    time: 25.2,
-                                    tape_nbr: 1,
-                                },
-                                {
-                                    id: 11310,
-                                    type: 'node',
-                                    children: [
-                                        {
-                                            type: 'leafe',
-                                            time: 1472.18,
-                                            tape_nbr: 1,
-                                        },
-                                        {
-                                            type: 'leafe',
-                                            time: 2103.32,
-                                            tape_nbr: 1,
-                                        },
-                                    ],
-                                },
-                            ],
-                        },
-                    ],
-                },
-                segments: {
-                    1: {
-                        199498: {
-                            id: 199498,
-                            type: 'Segment',
-                            text: {
-                                de: 'dummy',
-                                'de-public': 'dummy',
-                                ru: 'dummy',
-                                'ru-public': 'dummy',
-                            },
-                        },
-                    },
-                },
-                headings: {
-                    0: {
-                        id: 19499,
-                        type: 'Segment',
-                    },
-                },
+            {
+                type: 'sound',
+                contributions: [sound1, sound2],
             },
-        },
-        languages: {
-            2: {
+        ];
+        expect(actual).toEqual(expected);
+    });
+
+    test('also returns quality managers when edit view is true', () => {
+        const _state = dotProp.set(contributionState, 'archive.editView', true);
+
+        const actual = getGroupedContributions(_state);
+        const expected = [
+            {
+                type: 'interviewer',
+                contributions: [interviewer],
+            },
+            {
+                type: 'sound',
+                contributions: [sound1, sound2],
+            },
+            {
+                type: 'quality_manager_transcription',
+                contributions: [qualityManager],
+            },
+        ];
+        expect(actual).toEqual(expected);
+    });
+});
+
+/** Project Tests */
+
+describe('getCurrentProject', () => {
+    test("gets currently selected project's object", () => {
+        expect(getCurrentProject(projectState)).toEqual(
+            projectState.data.projects[1]
+        );
+    });
+
+    test('returns null if project cannot be found', () => {
+        const state = {
+            archive: {
+                projectId: 'cdoh',
+            },
+            data: {
+                projects: {},
+            },
+        };
+
+        expect(getCurrentProject(state)).toBeNull();
+    });
+});
+
+describe('getShowFeaturedInterviews', () => {
+    test('is false for mog project', () => {
+        const _state = dotProp.set(projectState, 'archive.projectId', 'mog');
+
+        expect(getShowFeaturedInterviews(_state)).toBeFalsy();
+    });
+
+    test('is false for campscapes project', () => {
+        const _state = dotProp.set(
+            projectState,
+            'archive.projectId',
+            'campscapes'
+        );
+        expect(getShowFeaturedInterviews(_state)).toBeFalsy();
+    });
+
+    test('is true otherwise', () => {
+        expect(getShowFeaturedInterviews(projectState)).toBeTruthy();
+    });
+});
+
+describe('getshowStartPageVideo', () => {
+    test('is true for mog project', () => {
+        const _state = dotProp.set(projectState, 'archive.projectId', 'mog');
+
+        expect(getShowStartPageVideo(_state)).toBeTruthy();
+    });
+
+    test('is false otherwise', () => {
+        expect(getShowStartPageVideo(projectState)).toBeFalsy();
+    });
+});
+
+test('getProjectTranslation gets project translation for current locale', () => {
+    const actual = getProjectTranslation(projectState);
+    const expected = projectState.data.projects[1].translations_attributes[0];
+    expect(actual).toEqual(expected);
+});
+
+describe('getIsCampscapesProject', () => {
+    test('is true when set to campscapes', () => {
+        const _state = dotProp.set(
+            projectState,
+            'archive.projectId',
+            'campscapes'
+        );
+
+        expect(getIsCampscapesProject(_state)).toBeTruthy();
+    });
+
+    test('is false otherwise', () => {
+        expect(getIsCampscapesProject(projectState)).toBeFalsy();
+    });
+});
+
+describe('getIsCatalog', () => {
+    test('returns is_catalog value', () => {
+        expect(getIsCatalog(projectState)).toEqual(
+            projectState.data.projects[1].is_catalog
+        );
+    });
+
+    test('is false otherwise', () => {
+        const _state = dotProp.set(
+            projectState,
+            'data.projects.1.is_catalog',
+            null
+        );
+
+        expect(getIsCatalog(_state)).toBe(false);
+    });
+});
+
+describe('getMapSections', () => {
+    test('returns map sections as ordered array', () => {
+        const actual = getMapSections(projectState);
+        const expected = [
+            {
                 id: 2,
-                type: 'Language',
+                type: 'MapSection',
+                name: 'usa',
+                order: 0,
             },
-        },
-        registry_entries: {
-            1: {
+            {
                 id: 1,
+                type: 'MapSection',
+                name: 'europe',
+                order: 1,
             },
-            23: {
-                id: 23,
-                type: 'RegistryEntry',
-                name: {
-                    de: 'Russland',
-                    en: 'Russia',
-                },
-                children_count: 77,
-            },
-        },
-        permissions: {
-            92: {
-                id: 92,
-                type: 'Permission',
-            },
-        },
-        segments: {
-            99015: {
-                type: 'Segment',
-            },
-        },
-        tasks: {
-            35: {
-                id: 35,
-                type: 'Task',
-            },
-        },
-        random_featured_interviews: {
-            cd009: {
-                id: 23,
-                type: 'Interview',
-            },
-        },
-    },
-};
+        ];
+        expect(actual).toEqual(expected);
+    });
 
-test('getLanguages gets languages object', () => {
-    expect(getLanguages(state)).toEqual(state.data.languages);
+    test('returns default section if map sections are blank', () => {
+        const _state = dotProp.set(
+            projectState,
+            'data.projects.1.map_sections',
+            {}
+        );
+        const actual = getMapSections(_state);
+        const expected = [DEFAULT_MAP_SECTION];
+        expect(actual).toEqual(expected);
+    });
+
+    test('returns default section if map sections is undefined', () => {
+        const _state = dotProp.set(
+            projectState,
+            'data.projects.1.map_sections',
+            undefined
+        );
+        const actual = getMapSections(_state);
+        const expected = [DEFAULT_MAP_SECTION];
+        expect(actual).toEqual(expected);
+    });
+});
+
+describe('getProjectLocales', () => {
+    test('gets project locales', () => {
+        expect(getProjectLocales(projectState)).toEqual(
+            projectState.data.projects[1].available_locales
+        );
+    });
+
+    test('gets default locales if no project is selected', () => {
+        const _state = dotProp.set(projectState, 'archive.projectId', null);
+        expect(getProjectLocales(_state)).toEqual(DEFAULT_LOCALES);
+    });
 });
 
 test('getCollectionsForCurrentProject gets collections object', () => {
-    expect(getCollectionsForCurrentProject(state)).toEqual(
-        state.data.projects[1].collections
+    expect(getCollectionsForCurrentProject(projectState)).toEqual(
+        projectState.data.projects[1].collections
     );
 });
 
 test('getTaskTypesForCurrentProject gets task_types object', () => {
-    expect(getTaskTypesForCurrentProject(state)).toEqual(
-        state.data.projects[1].task_types
+    expect(getTaskTypesForCurrentProject(projectState)).toEqual(
+        projectState.data.projects[1].task_types
     );
 });
 
 test('getRolesForCurrentProject gets roles object', () => {
-    expect(getRolesForCurrentProject(state)).toEqual(
-        state.data.projects[1].roles
+    expect(getRolesForCurrentProject(projectState)).toEqual(
+        projectState.data.projects[1].roles
     );
 });
 
-test('getStatuses gets statuses object', () => {
-    expect(getStatuses(state)).toEqual(state.data.statuses);
+test('getMediaStreamsForCurrentProject retrieves media streams', () => {
+    expect(getMediaStreamsForCurrentProject(projectState)).toEqual(
+        projectState.data.projects[1].media_streams
+    );
 });
+
+test('getContributionTypesForCurrentProject retrieves contributionTypes object', () => {
+    expect(getContributionTypesForCurrentProject(projectState)).toEqual(
+        projectState.data.projects[1].contribution_types
+    );
+});
+
+/** Registry Tests */
+test('getRegistryEntriesStatus gets registry entries status object', () => {
+    expect(getRegistryEntriesStatus(state)).toEqual(
+        state.data.statuses.registry_entries
+    );
+});
+
+test('getRegistryNameTypesForCurrentProject gets registry name types object', () => {
+    expect(getRegistryNameTypesForCurrentProject(state)).toEqual(
+        state.data.registry_name_types
+    );
+});
+
+/** Status Tests */
 
 test('getUsersStatus gets users status object', () => {
     expect(getUsersStatus(state)).toEqual(state.data.statuses.users);
@@ -349,10 +383,6 @@ test('getHeadingsStatus gets headings status object', () => {
 
 test('getHeadingsFetched gets if headings of current interview are fetched', () => {
     expect(getHeadingsFetched(state)).toBeTruthy();
-});
-
-test('getHeadings gets headings object of current interview', () => {
-    expect(getHeadings(state)).toEqual(state.data.interviews.cd003.headings);
 });
 
 test('getLanguagesStatus gets languages status object', () => {
@@ -405,12 +435,6 @@ describe('getCurrentRefTreeStatus', () => {
     });
 });
 
-test('getRegistryEntriesStatus gets registry entries status object', () => {
-    expect(getRegistryEntriesStatus(state)).toEqual(
-        state.data.statuses.registry_entries
-    );
-});
-
 test('getRolesStatus gets roles status object', () => {
     expect(getRolesStatus(state)).toEqual(state.data.statuses.roles);
 });
@@ -433,137 +457,14 @@ test('getTaskTypesStatus gets task types status object', () => {
     expect(getTaskTypesStatus(state)).toEqual(state.data.statuses.task_types);
 });
 
-test('getUsersStatus gets user registrations status object', () => {
-    expect(getUsersStatus(state)).toEqual(state.data.statuses.users);
-});
-
-test('getUsers gets users object', () => {
-    expect(getUsers(state)).toEqual(state.data.users);
-});
+/** User Tests */
 
 test('getCurrentUser gets user object of current user', () => {
     expect(getCurrentUser(state)).toEqual(state.data.users.current);
 });
 
-test('getPermissions gets permissions object', () => {
-    expect(getPermissions(state)).toEqual(state.data.permissions);
-});
-
-test('getRegistryEntries gets registry entries object', () => {
-    expect(getRegistryEntries(state)).toEqual(state.data.registry_entries);
-});
-
-test('getRegistryNameTypesForCurrentProject gets registry name types object', () => {
-    expect(getRegistryNameTypesForCurrentProject(state)).toEqual(
-        state.data.registry_name_types
-    );
-});
-
-test('getSegments gets segments object', () => {
-    expect(getSegments(state)).toEqual(state.data.segments);
-});
-
-test('getTasks gets tasks object', () => {
-    expect(getTasks(state)).toEqual(state.data.tasks);
-});
-
-test('getTaskTypesForCurrentProject gets task types object', () => {
-    expect(getTaskTypesForCurrentProject(state)).toEqual(
-        state.data.projects[1].task_types
-    );
-});
-
-test('getUsers gets user registrations object', () => {
-    expect(getUsers(state)).toEqual(state.data.users);
-});
-
-test('getRandomFeaturedInterviews gets featured interviews object', () => {
-    expect(getRandomFeaturedInterviews(state)).toEqual(
-        state.data.random_featured_interviews
-    );
-});
-
 test('getCurrentUserIsAdmin gets admin status of current user', () => {
     expect(getCurrentUserIsAdmin(state)).toEqual(
         state.data.users.current.admin
-    );
-});
-
-describe('getProjectLocales', () => {
-    test('gets project locales', () => {
-        expect(getProjectLocales(state)).toEqual(
-            state.data.projects[1].available_locales
-        );
-    });
-
-    test('gets default locales if no project is selected', () => {
-        const _state = dotProp.set(state, 'archive.projectId', null);
-        expect(getProjectLocales(_state)).toEqual(DEFAULT_LOCALES);
-    });
-});
-
-test('getCurrentInterviewFetched retrieves if current interview has been fetched', () => {
-    expect(getCurrentInterviewFetched(state)).toBe(true);
-});
-
-test('getCurrentRefTree retrieves ref tree of current interview', () => {
-    expect(getCurrentRefTree(state)).toEqual(
-        state.data.interviews.cd003.ref_tree
-    );
-});
-
-test('getFlattenedRefTree retrieves flattened version of ref tree', () => {
-    const actual = getFlattenedRefTree(state);
-    const expected = {
-        11310: {
-            id: 11310,
-            type: 'node',
-            children: [
-                {
-                    type: 'leafe',
-                    time: 1472.18,
-                    tape_nbr: 1,
-                },
-                {
-                    type: 'leafe',
-                    time: 2103.32,
-                    tape_nbr: 1,
-                },
-            ],
-        },
-        11786: {
-            id: 11786,
-            type: 'node',
-            children: [
-                {
-                    type: 'leafe',
-                    time: 25.2,
-                    tape_nbr: 1,
-                },
-            ],
-        },
-    };
-
-    expect(actual).toEqual(expected);
-});
-
-describe('getTranscriptFetched', () => {
-    test('returns true if transcript has been fetched', () => {
-        expect(getTranscriptFetched(state)).toBeTruthy();
-    });
-
-    test('returns false if transcript has not been fetched', () => {
-        const _state = dotProp.set(state, 'archive.archiveId', 'za085');
-        expect(getTranscriptFetched(_state)).toBeFalsy();
-    });
-});
-
-test('getContributorsFetched retrieves if contributors for current interview have been fetched', () => {
-    expect(getContributorsFetched(state)).toBe(true);
-});
-
-test('getMediaStreamsForCurrentProject retrieves media streams', () => {
-    expect(getMediaStreamsForCurrentProject(state)).toEqual(
-        state.data.projects[1].media_streams
     );
 });
