+++ app/javascript/modules/data/redux/selectors/index.js
@@ -1,4 +1,2 @@
 export * from './baseSelectors';
 export * from './dataSelectors';
-export * from './getGroupedContributions';
-export * from './projectSelectors';
