+++ app/javascript/modules/data/redux/selectors/utils.js
@@ -1,0 +1,8 @@
+export function projectByDomain(projects) {
+    return (
+        projects &&
+        Object.values(projects).find(
+            (project) => project.archive_domain === window.location.origin
+        )
+    );
+}
